Extracted chart data options to separate constants and moved DemoChart component to proper directory
